More workarounds for old glibc bug with strtoul.
Also fix a bug in the previous workaround.

+++ ParameterInfo.h
@@ -43,7 +43,7 @@ inline string inputOneValue <string> (istringstream &streamIn) {
 // in case of older versions of libc.
 template <>
 inline uint inputOneValue <uint> (istringstream &streamIn) {
-    int oneV;
+    int64_t oneV;
     streamIn >> oneV;
     return static_cast<uint>(oneV);
 };
